+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Display reposts made by current user in home timeline (reverting change made in v3.4.0).
 - Allow `proof.verificationMethod` to be DID URL.
 - Write message to log if FEP-1b12 activity is not supported.
+- Don't fetch embedded FEP-1b12 activity if it has same origin as its parent.
 
 ## [3.5.0] - 2024-09-17